# openapi.yaml
openapi: 3.0.3
info:
  title: Lumbung Mesari API
  description: |
    REST API for Lumbung Mesari Savings and Loans Cooperative System

    ## Features
    - JWT-based authentication with refresh tokens
    - Role-based access control (Member, Administrator, Super Administrator)
    - User management and registration
    - Savings management with transaction tracking
    - Loan management with installment scheduling
    - Expense management with approval workflow
    - Treasury system for cooperative balance management

    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Base URL
    All endpoints are prefixed with `/api`
  version: 1.0.0
  contact:
    name: Lumbung Mesari Development Team
    email: dev@lumbungmesari.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.lumbungmesari.com/api
    description: Production server
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with email or username and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              email_login:
                summary: Login with email
                value:
                  identifier: admin@lumbungmesari.com
                  password: admin123
              username_login:
                summary: Login with username
                value:
                  identifier: admin
                  password: admin123
      responses:
        '201':
          description: Successfully logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Register a new member account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Bad request - validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get new access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  description: Valid refresh token
              required:
                - refresh_token
      responses:
        '201':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      description: Retrieve authenticated user's profile information
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # User Management Endpoints
  /users:
    get:
      tags:
        - Users
      summary: Get users with filtering and pagination
      description: Retrieve list of users with role filtering and pagination (admin only)
      parameters:
        - name: role
          in: query
          description: Filter by user role
          schema:
            type: string
            enum: [member, administrator, superadministrator]
          example: member
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          example: 10
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Expense Management Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required:
        - identifier
        - password
      properties:
        identifier:
          type: string
          description: Email or username
          example: admin@lumbungmesari.com
        password:
          type: string
          description: User password
          example: admin123
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        password:
          type: string
          minLength: 6
          description: User password (minimum 6 characters)
          example: password123
        firstName:
          type: string
          description: User first name
          example: John
        lastName:
          type: string
          description: User last name
          example: Doe
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          description: JWT refresh token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        _links:
          type: object
          properties:
            self:
              type: object
              properties:
                href:
                  type: string
                  example: /api/auth/profile
            refresh:
              type: object
              properties:
                href:
                  type: string
                  example: /api/auth/refresh

    # User Schemas
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User ID
        email:
          type: string
          format: email
          description: User email
        fullname:
          type: string
          description: Full name of the user
          example: John Doe
        username:
          type: string
          description: Username
        phone_number:
          type: string
          description: Phone number
          example: '+628123456789'
        address:
          type: string
          description: User address
          example: 'Jl. Merdeka No. 123, Jakarta'
        status:
          type: string
          enum: [waiting_deposit, active, inactive, suspended]
          description: User status
    UserProfile:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            balance:
              type: number
              format: decimal
              description: Current balance
            totalSavings:
              type: number
              format: decimal
              description: Total savings amount
            activeLoans:
              type: integer
              description: Number of active loans

    # Expense Schemas
...
    UsersListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
          description: Array of user objects
        page:
          type: integer
          description: Current page number
          example: 1
        limit:
          type: integer
          description: Items per page
          example: 10
        next:
          type: boolean
          description: Whether there is a next page
          example: true
        prev:
          type: boolean
          description: Whether there is a previous page
          example: false

    # Common Schemas
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Error message
        error:
          type: string
          description: Error type
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
    ValidationErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          description: HTTP status code
          example: 400
        message:
          type: array
          items:
            type: string
          description: Validation error messages
          example:
            [
              'email must be a valid email',
              'password must be longer than or equal to 6 characters'
            ]
        error:
          type: string
          description: Error type
          example: Bad Request
        timestamp:
          type: string
          format: date-time
          description: Error timestamp